+++ MonsterChase/player.h
@@ -16,7 +16,7 @@ namespace MonsterChase
 	{
 	public:
 		Player();
-		Player(const Engine::TString & name, const Engine::Point2D<int> & i_position, int health);
+		Player(const Engine::TString & name, const Engine::Point2D<float> & i_position, int health);
 		~Player();
 		void PrintOutInfo() override;
